requirements.txt
altair==4.2.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
easyocr==1.7.0
entrypoints==0.4
filelock==3.12.0
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.30.0
importlib-metadata==6.6.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
Menu==3.2.2
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.24.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
option==2.1.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
platformdirs==3.5.1
protobuf==3.20.3
pyarrow==12.0.0
pyclipper==1.3.0.post4
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytesseract=0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
rich==13.3.5
scikit-image==0.20.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
smmap==5.0.0
streamlit==1.22.0
streamlit-option-menu==0.3.5
sympy==1.12
tenacity==8.2.2
tifffile==2023.4.12
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
validators==0.20.0
virtualenv==20.23.0
watchdog==3.0.0
zipp==3.15.0
